Added missing endpoints
