+++ pom.xml
@@ -35,7 +35,7 @@
             <manifest>
               <addClasspath>true</addClasspath>
               <classpathPrefix>lib/</classpathPrefix>
-              <mainClass>io.fi0x.Main</mainClass>
+              <mainClass>io.fi0x.languagegenerator.logic.Main</mainClass>
             </manifest>
           </archive>
         </configuration>
@@ -64,7 +64,7 @@
         <configuration>
           <archive>
             <manifest>
-              <mainClass>io.fi0x.Main</mainClass>
+              <mainClass>io.fi0x.languagegenerator.logic.Main</mainClass>
             </manifest>
           </archive>
           <descriptorRefs>
@@ -86,7 +86,7 @@
         <artifactId>javafx-maven-plugin</artifactId>
         <version>0.0.8</version>
         <configuration>
-          <mainClass>io.fi0x.Main</mainClass>
+          <mainClass>io.fi0x.languagegenerator.logic.Main</mainClass>
         </configuration>
       </plugin>
       <plugin>
@@ -108,7 +108,7 @@
               <stayAlive>true</stayAlive>
               <icon>src\main\resources\images\ico.ico</icon>
               <classPath>
-                <mainClass>io.fi0x.Main</mainClass>
+                <mainClass>io.fi0x.languagegenerator.logic.Main</mainClass>
               </classPath>
               <jre>
                 <path>jre</path>
