+++ src/main/resources/application.properties
@@ -1,3 +1,7 @@
 spring.mvc.view.prefix=/WEB-INF/jsp/
 spring.mvc.view.suffix=.jsp
+
 server.port=2345
+
+logging.level.org.springframework=info
+logging.level.io.fi0x.languagegenerator=info
